comment index, create TitleAnimation for TitleText to make TitleText.tsx more DRY
