+++ components/TitleText.tsx
@@ -1,28 +1,12 @@
 import { motion, AnimatePresence } from "framer-motion";
 import { css } from "styled-jsx/css";
 import Head from "next/head";
+import TitleAnimation from './TitleAnimation';
 
 interface TitleTextProps {
   titleText: string;
   onHelloPage: boolean;
 }
-
-// control styles for motion.div's
-const { className, styles } = css.resolve`
-  div {
-    display: block;
-    position: relative;
-    z-index: 1;
-    padding: 0;
-    margin: 0;
-  }
-`;
-
-const variants = {
-  moveUp: { opacity: 0, y: "0px", x: "-100px" },
-  center: { opacity: 1, y: "0", x: "0px" },
-  moveRight: { opacity: 0, y: "-100px", x: "0px" },
-};
 
 const TitleText: React.FC<TitleTextProps> = ({ onHelloPage, titleText }) => {
   return (
@@ -33,34 +17,20 @@ const TitleText: React.FC<TitleTextProps> = ({ onHelloPage, titleText }) => {
           rel="stylesheet"
         />
       </Head>
+      {/* Animate "HELLO" as it enters/leaves the DOM */}
       <AnimatePresence initial={false}>
         {onHelloPage && (
-          <motion.div
-            className={className}
-            initial="moveUp"
-            animate="center"
-            exit="moveRight"
-            transition={{ duration: 1 }}
-            variants={variants}
-          >
+          <TitleAnimation>
             <h1 className="title blue">{titleText}</h1>
-            {styles}
-          </motion.div>
+          </TitleAnimation>
         )}
       </AnimatePresence>
+      {/* Animate "ABOUT" as it enters/leaves the DOM  */}
       <AnimatePresence initial={false}>
         {!onHelloPage && (
-          <motion.div
-            className={className}
-            initial="moveUp"
-            animate="center"
-            exit="moveRight"
-            transition={{ duration: 1 }}
-            variants={variants}
-          >
+          <TitleAnimation>
             <h1 className="title red">{titleText}</h1>
-            {styles}
-          </motion.div>
+          </TitleAnimation>
         )}
       </AnimatePresence>
